more type-safety
