+++ deployables/app/app/balances/server/getTokenDetails.ts
@@ -1,14 +1,20 @@
+import type { HexAddress } from '@zodiac/schema'
 import { tokenSchema, type Chain, type TokenBalance } from '../types'
 import { api } from './api'
 
+type GetByAddressOptions = { address: HexAddress }
+type GetBySymbolOptions = { symbol: string }
+
+type GetTokenDetailOptions = GetByAddressOptions | GetBySymbolOptions
+
 export const getTokenDetails = async (
   chain: Chain,
-  address: string,
+  options: GetTokenDetailOptions,
 ): Promise<TokenBalance> => {
   const rawData = await api('/token', {
     schema: tokenSchema,
     data: {
-      id: address,
+      id: 'address' in options ? options.address : options.symbol,
       chain_id: chain.id,
     },
   })
